修改登录和注册

+++ src/main/java/com/chasel/blog/dao/IBaseDao.java
@@ -17,11 +17,11 @@ public interface IBaseDao<T> {
 
 	public void save(T t) throws BlogException;
 
-	public void delete(@Param("id") int id) throws BlogException;
+	public void delete(@Param("id") long id) throws BlogException;
 
 	public void update(T t) throws BlogException;
 
 	public List<T> findAll(T t) throws BlogException;
 
-	public T findById(@Param("id") int id) throws BlogException;
+	public T findById(@Param("id") long id) throws BlogException;
 }
